requirements.txt
# Misskey API
Misskey.py==4.1.0

# Gemini API (新ライブラリ)
google-genai

# データベース
aiosqlite

# スケジューラー
APScheduler

# タイムゾーン
pytz

# WebSocket
websockets

# YAML設定ファイル
PyYAML

# 設定管理
pydantic
pydantic-settings
